+++ data/src/main/kotlin/com/wepli/data/di/datasource/DataSourceModule.kt
@@ -1,4 +1,4 @@
-package com.wepli.data.di
+package com.wepli.data.di.datasource
 
 import com.wepli.data.applemusic.datasource.AppleMusicDataSource
 import com.wepli.data.applemusic.datasource.AppleMusicDataSourceImpl
@@ -23,6 +23,8 @@ import com.wepli.data.relaylist.datasource.remote.RelaylistDataSourceImpl
 import com.wepli.data.relaylist.datasource.remote.RelaylistSupabaseDataSourceImpl
 import com.wepli.data.song.datasource.SongDataSource
 import com.wepli.data.song.datasource.SongSupabaseDataSourceImpl
+import com.wepli.data.supabase.bucket.datasource.SupabaseBucketDataSource
+import com.wepli.data.supabase.bucket.datasource.SupabaseBucketDataSourceImpl
 import com.wepli.data.user.datasource.UserSupabaseDataSource
 import com.wepli.data.user.datasource.UserSupabaseDataSourceImpl
 import dagger.Binds
@@ -61,42 +63,4 @@ interface DataSourceModule {
     @Binds
     @Singleton
     fun bindDataStoreDataSource(dataStoreDataSourceImpl: DataStorePrefDataSourceImpl): DataStorePrefDataSource
-
-    /**
-     * Supabase
-     */
-    @Binds
-    @Singleton
-    @SupabaseDataSource
-    fun bindUserSupabaseDataSource(userSupabaseDataSource: UserSupabaseDataSourceImpl): UserSupabaseDataSource
-
-    @Binds
-    @Singleton
-    @SupabaseDataSource
-    fun bindPostSupabaseDataSource(postSupabaseDataSourceImpl: PostSupabaseDataSourceImpl): PostDataSource
-
-    @Binds
-    @Singleton
-    @SupabaseDataSource
-    fun bindSongSupabaseDataSource(songSupabaseDataSourceImpl: SongSupabaseDataSourceImpl): SongDataSource
-
-    @Binds
-    @Singleton
-    @SupabaseDataSource
-    fun bindRelaylistSupabaseDataSource(relaylistSupabaseDataSourceImpl: RelaylistSupabaseDataSourceImpl): RelaylistDataSource
-
-    @Binds
-    @Singleton
-    @SupabaseDataSource
-    fun bindPlaylistSupabaseDataSource(playlistDataSourceImpl: PlaylistSupabaseDataSourceImpl): PlaylistDataSource
-
-    @Binds
-    @Singleton
-    @SupabaseDataSource
-    fun bindKeywordSupabaseDataSource(keywordDataSourceImpl: KeywordSupabaseDatasourceImpl): KeywordDatasource
-
-    @Binds
-    @Singleton
-    @SupabaseDataSource
-    fun bindChartSupabaseDataSource(chartDataSourceImpl: ChartSupabaseDataSourceImpl): ChartDataSource
 }
