+++ data/src/main/kotlin/com/wepli/data/di/datastore/DataStoreModule.kt
@@ -1,4 +1,4 @@
-package com.wepli.data.di
+package com.wepli.data.di.datastore
 
 import android.content.Context
 import androidx.datastore.core.DataStore
